# generatedOpenapi.yaml
openapi: 3.1.0
info:
  title: Bank API
  description: "REST API for a Eagle Bank which allows a user to create, fetch, update\
    \ and delete a bank account and deposit or withdraw money from the account."
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local Development Server
...
paths:
...
  /v1/accounts/{accountNumber}/transactions:
    get:
      tags:
      - transaction
      summary: List transactions
      description: List transactions
      operationId: listAccountTransaction
      parameters:
      - name: accountNumber
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The list of transaction details
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ListTransactionsResponseDto"
        "400":
          description: The request didn't supply all the necessary data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponseDto"
        "401":
          description: Access token is missing or invalid
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: The user is not allowed to access the transactions
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Bank account was not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: An unexpected error occurred
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - bearerAuth: []
    post:
      tags:
      - transaction
      summary: Create a transaction
      description: Create a transaction
      operationId: createTransaction
      parameters:
      - name: accountNumber
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequestDto"
        required: true
      responses:
        "201":
          description: Transaction has been created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionResponseDto"
        "400":
          description: Invalid details supplied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponseDto"
        "401":
          description: Access token is missing or invalid
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: The user is not allowed to delete the bank account details
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Bank account was not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "422":
          description: Insufficient funds to process transaction
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: An unexpected error occurred
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - bearerAuth: []
...
  /v1/accounts/{accountNumber}/transactions/{transactionId}:
    get:
      tags:
      - transaction
      summary: Fetch transaction by ID.
      description: Fetch transaction by ID.
      operationId: fetchAccountTransactionByID
      parameters:
      - name: accountNumber
        in: path
        required: true
        schema:
          type: string
      - name: transactionId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The transaction details
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionResponseDto"
        "400":
          description: The request didn't supply all the necessary data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponseDto"
        "401":
          description: Access token is missing or invalid
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: The user is not allowed to access the transaction
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Bank account was not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: An unexpected error occurred
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - bearerAuth: []
components:
  schemas:
...
    BadRequestErrorResponseDto:
      type: object
      properties:
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetailDto"
    ErrorDetailDto:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        type:
          type: string
    ErrorResponseDto:
      type: object
      properties:
        message:
          type: string
...
    CreateTransactionRequestDto:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        currency:
          type: string
          minLength: 1
        type:
          type: string
          enum:
          - DEPOSIT
          - WITHDRAWAL
        reference:
          type: string
          maxLength: 255
          minLength: 0
      required:
      - amount
      - currency
      - type
    TransactionResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        amount:
          type: number
        currency:
          type: string
        type:
          type: string
          enum:
          - DEPOSIT
          - WITHDRAWAL
        reference:
          type: string
        userId:
          type: integer
          format: int64
        createdTimestamp:
          type: string
          format: date-time
...
    ListTransactionsResponseDto:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionResponseDto"
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT Authentication token (Bearer Token)
      scheme: bearer
      bearerFormat: JWT